requirements.txt
certifi==2022.12.7
chardet==5.1.0
click==8.1.3
dominate==2.7.0
Flask==2.2.3
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Flask-Login==0.6.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
requests==2.28.2
SQLAlchemy==2.0.7
urllib3==1.26.15
visitor==0.1.3
Werkzeug==2.2.3
WTForms==3.0.1
python-dotenv==1.0.0
python-decouple==3.8
gunicorn==20.1.0
email-validator==1.3.1
mysqlclient==2.1.1
